feat: Deal ErrorCode 추가

+++ src/main/java/ssafy/ssafyhome/deal/application/request/PriceRange.java
@@ -1,4 +1,4 @@
 package ssafy.ssafyhome.deal.application.request;
 
-public record PriceRange(Long max, Long min) {
+public record PriceRange(Integer max, Integer min) {
 }
